// src/services/user.service.d.ts
import type { UserCreateDTOType } from "../types/UserType.ts";
import type { User } from "../../prisma/src/generated/prisma/client.ts";
declare class UserService {
    getUsers(): Promise<{
        id: string;
        name: string;
        surname: string;
        age: number;
        email: string;
        password: string;
        phone: string | null;
        gender: import("../../prisma/src/generated/prisma/enums.ts").GenderEnum | null;
        photo: string | null;
        role_id: number;
        region_id: number;
        city: string;
        account_type: import("../../prisma/src/generated/prisma/enums.ts").AccountTypeEnum;
        balance: number;
        currency: import("../../prisma/src/generated/prisma/enums.ts").CurrencyEnum;
        premium_since: Date | null;
        premium_until: Date | null;
        is_verified: boolean;
        is_active: boolean;
        is_deleted: boolean;
        created_at: Date;
        updated_at: Date;
    }[]>;
    getUser(id: string): Promise<{
        id: string;
        name: string;
        surname: string;
        age: number;
        email: string;
        password: string;
        phone: string | null;
        gender: import("../../prisma/src/generated/prisma/enums.ts").GenderEnum | null;
        photo: string | null;
        role_id: number;
        region_id: number;
        city: string;
        account_type: import("../../prisma/src/generated/prisma/enums.ts").AccountTypeEnum;
        balance: number;
        currency: import("../../prisma/src/generated/prisma/enums.ts").CurrencyEnum;
        premium_since: Date | null;
        premium_until: Date | null;
        is_verified: boolean;
        is_active: boolean;
        is_deleted: boolean;
        created_at: Date;
        updated_at: Date;
    }>;
    getUserByEmail(email: string): Promise<User | null>;
    createUser(userDTO: UserCreateDTOType, regionId: number): Promise<{
        id: string;
        name: string;
        surname: string;
        age: number;
        email: string;
        password: string;
        phone: string | null;
        gender: import("../../prisma/src/generated/prisma/enums.ts").GenderEnum | null;
        photo: string | null;
        role_id: number;
        region_id: number;
        city: string;
        account_type: import("../../prisma/src/generated/prisma/enums.ts").AccountTypeEnum;
        balance: number;
        currency: import("../../prisma/src/generated/prisma/enums.ts").CurrencyEnum;
        premium_since: Date | null;
        premium_until: Date | null;
        is_verified: boolean;
        is_active: boolean;
        is_deleted: boolean;
        created_at: Date;
        updated_at: Date;
    }>;
    deleteUser(id: string): Promise<{
        id: string;
        name: string;
        surname: string;
        age: number;
        email: string;
        password: string;
        phone: string | null;
        gender: import("../../prisma/src/generated/prisma/enums.ts").GenderEnum | null;
        photo: string | null;
        role_id: number;
        region_id: number;
        city: string;
        account_type: import("../../prisma/src/generated/prisma/enums.ts").AccountTypeEnum;
        balance: number;
        currency: import("../../prisma/src/generated/prisma/enums.ts").CurrencyEnum;
        premium_since: Date | null;
        premium_until: Date | null;
        is_verified: boolean;
        is_active: boolean;
        is_deleted: boolean;
        created_at: Date;
        updated_at: Date;
    }>;
}
export declare const userService: UserService;
export {};
//# sourceMappingURL=user.service.d.ts.map