// src/configs/configs.d.ts
type configsType = {
    APP_PORT: string;
    APP_HOST: string;
    JWT_ACCESS_SECRET: string;
    JWT_REFRESH_SECRET: string;
    JWT_ACTIVATE_SECRET: string;
    JWT_RECOVERY_SECRET: string;
    JWT_ACCESS_LIFETIME: any;
    JWT_REFRESH_LIFETIME: any;
    JWT_ACTIVATE_LIFETIME: any;
    JWT_RECOVERY_LIFETIME: any;
};
export declare const configs: configsType;
export {};
//# sourceMappingURL=configs.d.ts.map