// src/repository/user.repository.d.ts
declare class UserRepository {
    getUsers(): Promise<{
        id: string;
        name: string;
        surname: string;
        age: number;
        email: string;
        password: string;
// ...
    }[]>;
// ...
}
export declare const userRepository: UserRepository;